# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "n8n Library API",
    "description": "API for the n8n workflow template library, service desk, knowledge base, and monitoring.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
  "security": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Use your API key as the bearer token (format: n8nlib_...)"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Alternative: pass API key via X-API-Key header"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie (browser-based auth)"
      }
    },
    "schemas": {
...
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "key_prefix": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "editor", "viewer"] },
          "last_used_at": { "type": "string", "format": "date-time", "nullable": true },
          "expires_at": { "type": "string", "format": "date-time", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/templates": {
      "post": {
        "tags": ["Templates"],
        "summary": "Create template",
        "security": [{ "bearerAuth": [] }, { "sessionAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name", "workflow"], "properties": { "name": { "type": "string" }, "description": { "type": "string" }, "workflow": { "type": "object" }, "category_ids": { "type": "array", "items": { "type": "integer" } } } } } } },
        "responses": { "200": { "description": "Template created" } }
      }
    },
...
    "/api/mcp/servers": {
      "get": {
        "tags": ["MCP"],
        "summary": "List MCP servers",
        "security": [{ "bearerAuth": [] }, { "sessionAuth": [] }],
        "responses": { "200": { "description": "MCP server list with status" } }
      },
      "post": {
        "tags": ["MCP"],
        "summary": "Add MCP server",
        "security": [{ "bearerAuth": [] }, { "sessionAuth": [] }],
        "responses": { "200": { "description": "Server added" } }
      }
    },
...
    "/api/api-keys": {
      "get": {
        "tags": ["API Keys"],
        "summary": "List your API keys",
        "security": [{ "bearerAuth": [] }, { "sessionAuth": [] }],
        "responses": { "200": { "description": "API key list", "content": { "application/json": { "schema": { "type": "object", "properties": { "keys": { "type": "array", "items": { "$ref": "#/components/schemas/ApiKey" } } } } } } } }
      },
      "post": {
        "tags": ["API Keys"],
        "summary": "Create API key",
        "description": "Creates a new API key. The full key is returned ONLY in this response — store it securely.",
        "security": [{ "bearerAuth": [] }, { "sessionAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string", "example": "CI/CD Pipeline" }, "role": { "type": "string", "enum": ["admin", "editor", "viewer"], "default": "viewer" }, "expires_in": { "type": "integer", "description": "Days until expiry (omit for never)" } } } } } },
        "responses": { "200": { "description": "API key created (includes full key)", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/ApiKey" }, { "type": "object", "properties": { "key": { "type": "string", "example": "n8nlib_a1b2c3d4..." } } }] } } } } }
      }
    },
    "/api/api-keys/all": {
      "get": {
        "tags": ["API Keys"],
        "summary": "List all API keys (admin)",
        "security": [{ "bearerAuth": [] }, { "sessionAuth": [] }],
        "responses": { "200": { "description": "All API keys across all users" } }
      }
    },
    "/api/api-keys/{id}": {
      "put": {
        "tags": ["API Keys"],
        "summary": "Update API key",
        "security": [{ "bearerAuth": [] }, { "sessionAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Key updated" } }
      },
      "delete": {
        "tags": ["API Keys"],
        "summary": "Delete API key",
        "security": [{ "bearerAuth": [] }, { "sessionAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Key deleted" } }
      }
    }
  }
}